Moves data loading functions to new io module
- Need to convert funcs from methods to functions
- Need to update tests
