+++ captest/__init__.py
@@ -4,6 +4,12 @@ from captest import (
     columngroups,
 )
 
+from captest.io import(
+    load_data,
+    load_das,
+    load_pvsyst,
+)
+
 from ._version import get_versions
 __version__ = get_versions()['version']
 del get_versions
